feat(fields): field groups in the record create/edit dialog
The dialog was group-aware for ordering only — it shared the draft hook and
the baseline merge, then rendered a flat list through a SortableContext the
panel work had already thrown away twice.

Extract the panel's drag model into a shared FieldGroupList (section walk,
drop zones, insert lines, group highlight, collision detection with the
-end precedence rule and the empty-group filter, and the portalled overlay)
plus a useFieldGroupDnd hook over the draft mutators, then consume both from
the dialog. The panel's behaviour is unchanged: it passes the same handlers
and the same defaults.

Two FieldPanel problems the nesting exposes:

- The last-row border rule is a direct-child selector, so a group wrapper
  breaks it both ways. Add a rowBorders variant — auto (today, default) and
  managed, which reads a data-last-row marker the renderer stamps. CSS
  cannot express "no field-row later in document order" across parents.
- Indenting a grouped ROW would push its content column past the panel's
  single resize divider. Indent the label text instead, keeping the
  label/content boundary global.

saveDraft now resolves a boolean so the Create/Edit tab can abort the switch
when the write fails instead of re-snapshotting over unsaved work.
